backend/requirements.txt

flask==3.0.0
flask-cors==4.0.0
firebase-admin==6.4.0
python-dotenv==1.0.0
pillow==10.1.0
pytesseract==0.3.10
python-barcode==0.15.1
pyzbar==0.1.9
requests==2.31.0
pandas==2.1.3
gunicorn==21.2.0